requirements.txt
Flask==3.0.0
azureml-dataprep==4.12.6
pandas==1.5.3